generate_model_answer/requirements.txt
openai==1.40.0
tqdm==4.66.5
python-dotenv==1.0.1